+++ changelog.md
@@ -6,7 +6,9 @@
 ## 02-2024
 
 * [added] Icon showing the mediafile is representative
-* [fixed] The single species identified mediafiles are set representative
+* [fixed] The single species identified media files are set representative
+* [added] Buttons are now organized in workflow scheme
+* [added] Expected time of processing is now shown for taxon classification
 
 ## 01-2024